pipelines/azurepipeline/code/training-null/requirements.txt
Pillow
pathlib2
requests
azureml-sdk
scikit-learn
future
numpy
mlflow
matplotlib
azureml-mlflow
